consumer/requirements.txt
kafka-python==2.0.2
ddtrace==0.59.1
structlog==21.5.0